+++ analysis/src/commands.ml
@@ -1,13 +1,14 @@
-let completion ~debug ~source ~kind_file ~pos ~full =
+let completion ~state ~debug ~source ~kind_file ~pos ~full =
   match
-    Completions.get_completions ~debug ~source ~kind_file ~pos ~full
+    Completions.get_completions ~debug ~source ~kind_file ~pos ~full ~state
       ~for_hover:false
   with
   | None -> []
   | Some (completions, full, _) ->
-    completions |> List.map (Completion_back_end.completion_to_item ~full)
+    completions
+    |> List.map (Completion_back_end.completion_to_item ~state ~full)
 
-let completion_resolve ~(full : Shared_types.full option) ~module_path =
+let completion_resolve ~state ~(full : Shared_types.full option) ~module_path =
   (* We ignore the internal module path as of now because there's currently
      no use case for it. But, if we wanted to move resolving documentation
      for regular modules and not just file modules to the completionResolve
@@ -25,7 +26,9 @@ let completion_resolve ~(full : Shared_types.full option) ~module_path =
         Printf.printf "[completion_resolve] Could not load cmt\n";
       None
     | Some full -> (
-      match Process_cmt.file_for_module ~package:full.package module_name with
+      match
+        Process_cmt.file_for_module ~state ~package:full.package module_name
+      with
       | None ->
         if Debug.verbose () then
           Printf.printf "[completion_resolve] Did not find file for module %s\n"
@@ -41,7 +44,7 @@ let completion_resolve ~(full : Shared_types.full option) ~module_path =
          (Lsp.Types.MarkupContent.create ~kind:Lsp.Types.MarkupKind.Markdown
             ~value))
 
-let hover ~source ~kind_file ~pos ~supports_markdown_links ~full ~debug =
+let hover ~state ~source ~kind_file ~pos ~supports_markdown_links ~full ~debug =
   let result =
     match full with
     | None -> None
@@ -52,7 +55,7 @@ let hover ~source ~kind_file ~pos ~supports_markdown_links ~full ~debug =
           Printf.printf
             "Nothing at that position. Now trying to use completion.\n";
         match
-          Hover.get_hover_via_completions ~debug ~source ~kind_file ~pos
+          Hover.get_hover_via_completions ~debug ~source ~kind_file ~pos ~state
             ~for_hover:true ~supports_markdown_links ~full:(Some full)
         with
         | None -> None
@@ -63,7 +66,9 @@ let hover ~source ~kind_file ~pos ~supports_markdown_links ~full ~debug =
           | LModule _ | TopLevelModule _ -> true
           | TypeDefinition _ | Typed _ | Constant _ -> false
         in
-        let uri_loc_opt = References.definition_for_loc_item ~full loc_item in
+        let uri_loc_opt =
+          References.definition_for_loc_item ~state ~full loc_item
+        in
         let skip_zero =
           match uri_loc_opt with
           | None -> false
@@ -77,7 +82,7 @@ let hover ~source ~kind_file ~pos ~supports_markdown_links ~full ~debug =
             && pos_is_zero loc.loc_end
         in
         if skip_zero then None
-        else Hover.new_hover ~supports_markdown_links ~full loc_item)
+        else Hover.new_hover ~state ~supports_markdown_links ~full loc_item)
   in
   match result with
   | None -> None
@@ -90,12 +95,12 @@ let hover ~source ~kind_file ~pos ~supports_markdown_links ~full ~debug =
                  ~kind:Lsp.Types.MarkupKind.Markdown ~value))
          ())
 
-let signature_help ~source ~kind_file ~pos ~allow_for_constructor_payloads ~full
-    ~debug =
+let signature_help ~state ~source ~kind_file ~pos
+    ~allow_for_constructor_payloads ~full ~debug =
   Signature_help.signature_help ~debug ~source ~kind_file ~pos
-    ~allow_for_constructor_payloads ~full
+    ~allow_for_constructor_payloads ~full ~state
 
-let definition ~full ~pos ~debug =
+let definition ~state ~full ~pos ~debug =
   let location_opt =
     match full with
     | None -> None
@@ -103,7 +108,7 @@ let definition ~full ~pos ~debug =
       match References.get_loc_item ~full ~pos ~debug with
       | None -> None
       | Some loc_item -> (
-        match References.definition_for_loc_item ~full loc_item with
+        match References.definition_for_loc_item ~state ~full loc_item with
         | None -> None
         | Some (uri, loc) when not loc.loc_ghost ->
           let is_interface = full.file.uri |> Uri.is_interface in
@@ -130,7 +135,7 @@ let definition ~full ~pos ~debug =
   in
   location_opt
 
-let type_definition ~full ~pos ~debug =
+let type_definition ~state ~full ~pos ~debug =
   let maybe_location =
     match full with
     | None -> None
@@ -138,7 +143,7 @@ let type_definition ~full ~pos ~debug =
       match References.get_loc_item ~full ~pos ~debug with
       | None -> None
       | Some loc_item -> (
-        match References.type_definition_for_loc_item ~full loc_item with
+        match References.type_definition_for_loc_item ~state ~full loc_item with
         | None -> None
         | Some (uri, loc) ->
           Some
@@ -148,7 +153,7 @@ let type_definition ~full ~pos ~debug =
   in
   maybe_location
 
-let references ~full ~pos ~debug =
+let references ~state ~full ~pos ~debug =
   let all_locs =
     match full with
     | None -> []
@@ -157,7 +162,7 @@ let references ~full ~pos ~debug =
       | None -> []
       | Some loc_item ->
         let all_references =
-          References.all_references_for_loc_item ~full loc_item
+          References.all_references_for_loc_item ~state ~full loc_item
         in
         all_references
         |> List.fold_left
@@ -176,7 +181,7 @@ let references ~full ~pos ~debug =
   in
   all_locs
 
-let rename ~full ~pos ~new_name ~debug =
+let rename ~state ~full ~pos ~new_name ~debug =
   let result =
     match full with
     | None -> None
@@ -185,7 +190,7 @@ let rename ~full ~pos ~new_name ~debug =
       | None -> None
       | Some loc_item ->
         let all_references =
-          References.all_references_for_loc_item ~full loc_item
+          References.all_references_for_loc_item ~state ~full loc_item
         in
         let references_to_toplevel_modules =
           all_references